Add MONO_PATH when running C# executables under Linux or MacOS
C# executables do not have an rpath. Therefore, when building custom
targets or running tests, MONO_PATH has to be set on Linux or MacOS to
make sure that libraries are found. Under windows this is achieved by
setting PATH (which is used for binaries, native libraries and C#
libraries).