Add MONO_PATH or DEVPATH when running C# executables
C# executables do not have an rpath. Therefore, when building custom
targets or running tests, MONO_PATH has to be set on Linux or MacOS to
make sure that libraries are found. Under windows this is achieved by
setting DEVPATH. Note that on windows DEVPATH is only used when the
application config file sets developmentMode to true.